refactor: use subqueries for CTI
Using sub queries for Class Table Inheritance makes things simpler as there is no need to mess around with aliases, or do nasty things with entities after they have been hydrated, and there's no risk of joined tables to conflict with each other.

However, we must be careful that this refactor does not affect performances in a negative way.